Increased modularity of the classes page
